requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
numpy>=1.24.0
selenium>=4.15.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
lxml>=4.9.0
openai>=1.0.0
pytz>=2023.3